refactor out week as component
